+++ src/main/java/fr/raksrinana/fallingtree/config/ConfigCache.java
@@ -20,6 +20,7 @@ public class ConfigCache{
 	private Set<Block> leavesBlacklist;
 	private Set<Block> logsBlacklist;
 	private Set<Block> leavesWhitelist;
+	private Set<Block> leavesNonDecayWhitelist;
 	private Set<Block> logsWhitelist;
 	private Set<Block> adjacentBlocksWhitelist;
 	private Set<Block> adjacentBlocksBase;
@@ -29,6 +30,7 @@ public class ConfigCache{
 		toolsWhitelist = null;
 		leavesBlacklist = null;
 		leavesWhitelist = null;
+		leavesNonDecayWhitelist = null;
 		logsBlacklist = null;
 		logsWhitelist = null;
 		adjacentBlocksWhitelist = null;
@@ -70,13 +72,6 @@ public class ConfigCache{
 		return leavesWhitelist;
 	}
 	
-	public Collection<Block> getWhitelistedLogs(Supplier<Collection<String>> collectionSupplier){
-		if(Objects.isNull(logsWhitelist)){
-			logsWhitelist = getAsBlocks(collectionSupplier.get());
-		}
-		return logsWhitelist;
-	}
-	
 	public Collection<Block> getWhitelistedAdjacentBlocks(Supplier<Collection<String>> collectionSupplier){
 		if(Objects.isNull(adjacentBlocksWhitelist)){
 			adjacentBlocksWhitelist = getAsBlocks(collectionSupplier.get());
@@ -92,10 +87,25 @@ public class ConfigCache{
 			adjacentBlocksBase.addAll(BlockTags.LOGS.values());
 			adjacentBlocksBase.addAll(getWhitelistedLogs(FallingTree.config.getTreesConfiguration()::getWhitelistedLogsStr));
 			adjacentBlocksBase.addAll(getWhitelistedLeaves(FallingTree.config.getTreesConfiguration()::getWhitelistedLeavesStr));
+			adjacentBlocksBase.addAll(getWhitelistedNonDecayLeaves(FallingTree.config.getTreesConfiguration()::getWhitelistedNonDecayLeavesStr));
 			adjacentBlocksBase.removeAll(getBlacklistedLogs(FallingTree.config.getTreesConfiguration()::getBlacklistedLogsStr));
 			adjacentBlocksBase.removeAll(getBlacklistedLeaves(FallingTree.config.getTreesConfiguration()::getBlacklistedLeavesStr));
 		}
 		return adjacentBlocksBase;
+	}
+	
+	public Collection<Block> getWhitelistedLogs(Supplier<Collection<String>> collectionSupplier){
+		if(Objects.isNull(logsWhitelist)){
+			logsWhitelist = getAsBlocks(collectionSupplier.get());
+		}
+		return logsWhitelist;
+	}
+	
+	public Collection<Block> getWhitelistedNonDecayLeaves(Supplier<Collection<String>> collectionSupplier){
+		if(Objects.isNull(leavesNonDecayWhitelist)){
+			leavesNonDecayWhitelist = getAsBlocks(collectionSupplier.get());
+		}
+		return leavesNonDecayWhitelist;
 	}
 	
 	public static ConfigCache getInstance(){